chore(integrations/Intercom): Refactor intercom before implementing proactive interfaces (#14422)
