+++ integrations/intercom/.gitignore
@@ -1,1 +1,2 @@
 .botpress
+temp.ts
