+++ integrations/intercom/src/channels/index.ts
@@ -1,0 +1,6 @@
+import { channel } from './channel'
+import * as bp from '.botpress'
+
+export const channels = {
+  channel,
+} satisfies bp.IntegrationProps['channels']
