+++ integrations/intercom/src/misc/auth.ts
