+++ integrations/intercom/src/misc/types.ts
@@ -1,4 +1,6 @@
 import * as sdk from '@botpress/sdk'
+import { z } from '@botpress/sdk'
+import { conversationSourceSchema, webhookNotificationSchema } from 'src/misc/schemas'
 import * as bp from '.botpress'
 
 export type Handler = bp.IntegrationProps['handler']
@@ -20,3 +22,10 @@ export type MessageDefinition = sdk.MessageDefinition
 export type ActionProps = bp.AnyActionProps
 export type MessageHandlerProps = bp.AnyMessageProps
 export type AckFunction = bp.AnyAckFunction
+
+export type VerifyResult =
+  | { result: 'success'; isError: false; parsedNotification: z.infer<typeof webhookNotificationSchema> }
+  | { result: 'error'; isError: true; message: string }
+  | { result: 'ignore'; isError: false; message?: string }
+
+export type IntercomMessage = z.infer<typeof conversationSourceSchema>
